Refactor global styles and add instructions component
- Refactored the global styles in GlobalStyles.js to use styled-components for better maintainability and readability.
- Added a new Instructions component in Instructions.jsx to provide step-by-step instructions on how to use the Call Center Cost Estimator.
